separate weapons to different files by their type
