+++ src/npc/GoblinRogue.ts
@@ -2,26 +2,26 @@ import DoubleStrike from '../Ability/DoubleStrike';
 import AttributeType from '../Character/Attributes/AttributeType';
 import StatType from '../Character/Stats/StatType';
 import { armour } from '../Equipment/Armour';
-import { weapons } from '../Equipment/Weapon';
+import { weapons } from '../Equipment/Weapon/weapons';
 import NPC from './NPC';
 
 const GoblinRogue: NPC = {
     name: 'Goblin Rogue',
     attributes: {
-        [AttributeType.WeaponSkill]: {base: 0, perLvl: 1},
-        [AttributeType.Strength]: {base: 10, perLvl: 1},
-        [AttributeType.Dexterity]: {base: 5, perLvl: 1},
-        [AttributeType.Perception]: {base: 0, perLvl: 0.5},
-        [AttributeType.Constitution]: {base: 0, perLvl: 0.75},
+        [AttributeType.WeaponSkill]: { base: 0, perLvl: 1 },
+        [AttributeType.Strength]: { base: 10, perLvl: 1 },
+        [AttributeType.Dexterity]: { base: 5, perLvl: 1 },
+        [AttributeType.Perception]: { base: 0, perLvl: 0.5 },
+        [AttributeType.Constitution]: { base: 0, perLvl: 0.75 },
     },
     stats: {
         [StatType.MaxHealth]: { base: 12, perLvl: 3 },
         [StatType.Armour]: { base: 0, perLvl: 0.8 },
-    
+
         [StatType.Damage]: { base: 0, perLvl: 0.2 },
 
         [StatType.ArmourPenetration]: { base: 0, perLvl: 0.2 },
-    
+
         [StatType.ManaCost]: { base: 100 }
     },
     equipment: {
